% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\docType{data}
\name{.kariocas_internal_colors}
\alias{.kariocas_internal_colors}
\title{karioCaS Visualization Styles & Helpers}
\format{
An object of class \code{list} of length 5.
}
\usage{
.kariocas_internal_colors
}
\description{
Single Source of Truth for visual elements in karioCaS.
Follows Nature/Science publication standards.
}
\keyword{datasets}
