% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{get_kariocas_colors}
\alias{get_kariocas_colors}
\title{Get karioCaS Official Colors}
\usage{
get_kariocas_colors(
  type = c("ranks", "domains", "special", "upset", "heatmap")
)
}
\arguments{
\item{type}{Character. The palette to retrieve: "ranks", "domains", "special", "upset", or "heatmap".}
}
\value{
A named vector or list of colors.
}
\description{
Securely retrieves the official color palettes used across the karioCaS package.
}
