% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{get_kariocas_labels}
\alias{get_kariocas_labels}
\title{Standard Axis Labels with HTML/Markdown Styling
Uses ggtext syntax for mixing fonts and sizes.}
\usage{
get_kariocas_labels()
}
\description{
Standard Axis Labels with HTML/Markdown Styling
Uses ggtext syntax for mixing fonts and sizes.
}
