% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{get_kariocas_linetypes}
\alias{get_kariocas_linetypes}
\title{Get karioCaS Official Linetypes}
\usage{
get_kariocas_linetypes()
}
\value{
A named character vector of linetypes.
}
\description{
Retrieves standard linetypes for read and taxa retention plots.
}
