% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{get_kariocas_shapes}
\alias{get_kariocas_shapes}
\title{Get karioCaS Official Shapes}
\usage{
get_kariocas_shapes(type = "ranks")
}
\arguments{
\item{type}{Character. Currently supports "ranks".}
}
\value{
A named vector of shape integers.
}
\description{
Securely retrieves the official ggplot2 shapes for taxonomic ranks and special metrics.
}
