% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\docType{data}
\name{kariocas_dims}
\alias{kariocas_dims}
\title{Standard Output Dimensions (A4 Landscape)}
\format{
An object of class \code{list} of length 2.
}
\usage{
kariocas_dims
}
\description{
Standard Output Dimensions (A4 Landscape)
}
\keyword{datasets}
