% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\docType{data}
\name{kariocas_labels}
\alias{kariocas_labels}
\title{Standard Axis Labels with Mathematical Notation
Uses atop() for line breaks in expressions.}
\format{
An object of class \code{list} of length 3.
}
\usage{
kariocas_labels
}
\description{
Standard Axis Labels with Mathematical Notation
Uses atop() for line breaks in expressions.
}
\keyword{datasets}
