% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\docType{data}
\name{kariocas_linetypes}
\alias{kariocas_linetypes}
\title{Official karioCaS Line Types}
\format{
An object of class \code{character} of length 3.
}
\usage{
kariocas_linetypes
}
\description{
Official karioCaS Line Types
}
\keyword{datasets}
