% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{label_k_number}
\alias{label_k_number}
\title{Format numbers with K/M suffixes}
\usage{
label_k_number(x)
}
\arguments{
\item{x}{A numeric vector to format.}
}
\description{
Format numbers with K/M suffixes
}
