% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{label_kariocas_auto}
\alias{label_kariocas_auto}
\title{Smart Number Formatter for Logs
Displays integers as integers (100) and small fractions with decimals (0.01).}
\usage{
label_kariocas_auto(x)
}
\description{
Smart Number Formatter for Logs
Displays integers as integers (100) and small fractions with decimals (0.01).
}
