% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{plot_kariocas_empty}
\alias{plot_kariocas_empty}
\title{Generate a Standard "No Data" Plot (Skeleton Style)}
\usage{
plot_kariocas_empty(
  title_text = "No Data",
  subtitle_text = NULL,
  x_label = NULL,
  y_label = NULL
)
}
\description{
Generate a Standard "No Data" Plot (Skeleton Style)
}
