% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{scale_y_kariocas_log10}
\alias{scale_y_kariocas_log10}
\alias{scale_x_kariocas_log10}
\title{Log10 Scale with Smart Integer Formatting}
\usage{
scale_y_kariocas_log10(...)

scale_x_kariocas_log10(...)
}
\arguments{
\item{...}{Arguments passed to scale_y_continuous}
}
\description{
Log10 Scale with Smart Integer Formatting
}
