% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cariokas_002_upset_plot.R
\name{upset_karioCaS_never_are}
\alias{upset_karioCaS_never_are}
\title{Generate UpSet Plots per Sample/Domain/Level (Step 002)}
\usage{
upset_karioCaS_never_are(project_dir, import_script_path = NULL)
}
\arguments{
\item{project_dir}{Path to the project root.}

\item{import_script_path}{Deprecated.}
}
\description{
"karioCaS never are upset!"
}
