% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cariokas_000_mpa_import_reorganize_fuction.R
\name{import_mpa_data}
\alias{import_mpa_data}
\title{Import and Process MPA Data (Project Structure Based)}
\usage{
import_mpa_data(project_dir, log_filename = "log_000_data_import.txt")
}
\arguments{
\item{project_dir}{Path to the root of the project (e.g., "/Users/name/project_x").}

\item{log_filename}{Name of the log file (Default: "log_000_data_import.txt").}
}
\value{
A data.frame containing the processed count matrix.
}
\description{
This function enforces a specific project structure. It takes the project root directory,
looks specifically for a folder named "000_mpa_original", processes the .mpa files therein,
and saves logs to "log_archives".
}
