% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\docType{data}
\name{kariocas_shapes}
\alias{kariocas_shapes}
\title{Official karioCaS Shapes}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
kariocas_shapes
}
\description{
Official karioCaS Shapes
}
\keyword{datasets}
