% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{label_kariocas_auto}
\alias{label_kariocas_auto}
\title{Smart Number Formatter for Logs
Default uses comma for thousands to avoid conflict with decimal dot.}
\usage{
label_kariocas_auto(x)
}
\description{
Smart Number Formatter for Logs
Default uses comma for thousands to avoid conflict with decimal dot.
}
