% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{plot_kariocas_empty}
\alias{plot_kariocas_empty}
\title{Generate a Standard "No Data" Plot}
\usage{
plot_kariocas_empty(
  title_text = "No Data",
  subtitle_text = NULL,
  x_label = NULL,
  y_label = NULL
)
}
\arguments{
\item{title_text}{String for the plot title.}

\item{subtitle_text}{String for the plot subtitle.}

\item{x_label}{String for the x-axis label.}

\item{y_label}{String for the y-axis label.}
}
\description{
Generate a Standard "No Data" Plot
}
