% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_1000_reliable_mpa.R
\name{retrieve_selected_taxa}
\alias{retrieve_selected_taxa}
\title{Retrieve Selected Taxa with Domain-Specific Thresholds (Final Mosaic Step)}
\usage{
retrieve_selected_taxa(
  project_dir,
  tax_level = NULL,
  CS_A,
  reads_min_A,
  CS_B,
  reads_min_B,
  CS_E,
  reads_min_E,
  CS_V,
  reads_min_V
)
}
\arguments{
\item{project_dir}{Path to the project root.}

\item{tax_level}{Specific taxonomic level to filter (e.g., "Species", "Genus"). If NULL, keeps all levels.}

\item{CS_A}{Character/Numeric. CS suffix for Archaea (e.g., 06, 6, 60).}

\item{reads_min_A}{Integer. Min reads for Archaea.}

\item{CS_B}{Character/Numeric. CS suffix for Bacteria.}

\item{reads_min_B}{Integer. Min reads for Bacteria.}

\item{CS_E}{Character/Numeric. CS suffix for Eukaryota.}

\item{reads_min_E}{Integer. Min reads for Eukaryota.}

\item{CS_V}{Character/Numeric. CS suffix for Viruses.}

\item{reads_min_V}{Integer. Min reads for Viruses.}
}
\description{
This function creates a "biological mosaic" for each sample using 'karioCaS_TSE.rds'.
It applies smart matching for CS values (e.g., input 04 matches CS40) and generates
the final hybrid .mpa files with detailed logging.
ENFORCES strict > 0 filter to prevent zero-count taxa from appearing in output.
}
