% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_style.R
\name{scale_y_kariocas_log10}
\alias{scale_y_kariocas_log10}
\alias{scale_x_kariocas_log10}
\title{Log10 Scale with Smart Integer Formatting}
\usage{
scale_y_kariocas_log10(labels = label_kariocas_auto, ...)

scale_x_kariocas_log10(labels = label_kariocas_auto, ...)
}
\arguments{
\item{labels}{A formatting function or vector. Defaults to label_kariocas_auto.}

\item{...}{Other arguments passed to scale_y_continuous}
}
\description{
Log10 Scale with Smart Integer Formatting
}
