% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cariokas_001_taxa_retention.R
\name{taxa_retention}
\alias{taxa_retention}
\title{Run Confidence Score Retention Analysis (Step 001)}
\usage{
taxa_retention(project_dir, import_script_path = NULL)
}
\arguments{
\item{project_dir}{Root path of the project.}

\item{import_script_path}{Deprecated.}
}
\description{
Executes read retention analysis based on Confidence Score (Kraken/Bracken).
}
