% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kariocas_002_upset_plot.R
\name{upset_kariocas}
\alias{upset_kariocas}
\title{Generate UpSet Plots per Sample/Domain/Level (Step 002)}
\usage{
upset_kariocas(project_dir)
}
\arguments{
\item{project_dir}{Path to the project root.}
}
\value{
Generates PDF plots saved in the project directory and returns a data frame invisibly.
}
\description{
"karioCaS never are upset!"
Generates UpSet plots with detailed logging.
}
\examples{
# Get the path to the included toy dataset
toy_project <- system.file("extdata", "your_project_name", package = "karioCaS")

# upset_kariocas(project_dir = toy_project)
}
